migrated to navigation component and sign in activity moved to activity fragment
